requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
edge-tts>=7.2.7
PyYAML>=6.0.2
python-multipart>=0.0.20
